feat: add safety module
