+++ core/consensus/aura/primitives/Cargo.toml
@@ -12,6 +12,7 @@ runtime_primitives = { package = "sr-primitives", path = "../../../sr-primitives
 primitives = { package = "substrate-primitives", path = "../../../primitives", default-features = false }
 rstd = { package = "sr-std", path = "../../../sr-std", default-features = false }
 consensus = { package = "srml-consensus", path = "../../../../srml/consensus", default-features = false }
+safety-primitives = { package = "substrate-consensus-safety-primitives", path = "../../safety/primitives" }
 
 [features]
 default = ["std"]
