+++ core/consensus/safety/primitives/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "substrate-consensus-safety-primitives"
+version = "2.0.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Primitives for Accountable Safety"
+edition = "2018"
