+++ core/consensus/slots/src/aux_schema.rs
@@ -20,7 +20,7 @@ use codec::{Encode, Decode};
 use client::backend::AuxStore;
 use client::error::{Result as ClientResult, Error as ClientError};
 use runtime_primitives::traits::{Header, Verify};
-use consensus_srml::EquivocationProof;
+use safety_primitives::EquivocationProof;
 
 const SLOT_HEADER_MAP_KEY: &[u8] = b"slot_header_map";
 const SLOT_HEADER_START: &[u8] = b"slot_header_start";
